Create PopupDialog.cs
For those of you like myself who have struggled with modal popups in Xamarin Forms, this popup class could be the answer!  If you have worked with XLab’s PopupLayout, you will have learned that it has limitation.  It is not truly modal and it does not handle resizing and orientation changes.  Changing the orientation clears the content.  I have searched all over the Internet for a good easy to use popup and could not find one, so I decided to make my own.  This popup class handles all orientation change events.  It also has an optional modal parameter that makes it truly modal.  It scales from .25 to 1.  If the scale is set to 1, it will cover the content area of the screen like a sheet window.  Since it is a class you can freely manipulate the code to fit your needs.  It also has a closed event to track changes.  
This is how you call it. 
        PopupContent.ShowPopup(TableView, 1, modal: true, title: "Perfect Popup ");