Create HomePage.cs
This is a sample implementation of the custom Popup class from a Xamarin Forms Navigation page.  The popup can display any view that you pass it.
It should be called from the page's OnSizeAllocated event to handle all orientation changes.  The content page's root content needs to be set to the PopupContent variable like below.
Content = PopupContent;